requirements.txt
scipy<1.10.0
numpy
matplotlib
pandas<2.0.0
shapely>=1.7.0
geopandas
fiona>=1.9
contextily
xarray
dask
netcdf4>=1.5.3
bottleneck
xugrid>=0.6.1
cdsapi
pydap>=3.3.0
hydrolib-core>=0.5.1
meshkernel>=2.1.0
